refactor: updated openapi specs
