+++ api/openapispec/swagger.json
@@ -537,7 +537,7 @@
                     "application/json"
                 ],
                 "tags": [
-                    "resource"
+                    "insight"
                 ],
                 "summary": "GetDetail returns a Kubernetes resource by name, namespace, cluster, apiVersion and kind.",
                 "parameters": [
@@ -631,7 +631,7 @@
                     "application/json"
                 ],
                 "tags": [
-                    "resource"
+                    "insight"
                 ],
                 "summary": "GetEvents returns events for a Kubernetes resource by name, namespace, cluster, apiVersion and kind.",
                 "parameters": [
@@ -804,7 +804,7 @@
                     "application/json"
                 ],
                 "tags": [
-                    "resource"
+                    "insight"
                 ],
                 "summary": "Get returns a Kubernetes resource summary by name, namespace, cluster, apiVersion and kind.",
                 "parameters": [
@@ -892,7 +892,7 @@
                     "application/json"
                 ],
                 "tags": [
-                    "resource"
+                    "insight"
                 ],
                 "summary": "GetTopology returns a topology map for a Kubernetes resource by name, namespace, cluster, apiVersion and kind.",
                 "parameters": [
@@ -976,14 +976,14 @@
                 }
             }
         },
-        "/api/v1/resource/search": {
+        "/api/v1/search": {
             "get": {
                 "description": "This endpoint returns an array of Kubernetes runtime Object matched using the query from context.",
                 "produces": [
                     "application/json"
                 ],
                 "tags": [
-                    "resource"
+                    "search"
                 ],
                 "summary": "SearchForResource returns an array of Kubernetes runtime Object matched using the query from context.",
                 "parameters": [
