+++ pkg/core/handler/search/search.go
@@ -32,7 +32,7 @@ import (
 //
 //	@Summary		SearchForResource returns an array of Kubernetes runtime Object matched using the query from context.
 //	@Description	This endpoint returns an array of Kubernetes runtime Object matched using the query from context.
-//	@Tags			resource
+//	@Tags			search
 //	@Produce		json
 //	@Param			query		query		string			true	"The query to use for search. Required"
 //	@Param			pattern		query		string			true	"The search pattern. Can be either sql or dsl. Required"
@@ -45,7 +45,7 @@ import (
 //	@Failure		405			{string}	string			"Method Not Allowed"
 //	@Failure		429			{string}	string			"Too Many Requests"
 //	@Failure		500			{string}	string			"Internal Server Error"
-//	@Router			/api/v1/resource/search [get]
+//	@Router			/api/v1/search [get]
 func SearchForResource(resourceMgr *resource.ResourceManager, searchStorage storage.SearchStorage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// Extract the context and logger from the request.
